Add a hook on theme load
The first use is to set the forground used for the hidden leading
stars in org  mode so that it matches the background of the default
face, effectively making them invisible.

Later we'll use it for writing a ~/.Xdefaults file so that the
background matches the theme background, eliminating the flashing at
startup.